refactor: state, event, action 별도 파일로 분리
