+++ feature/book/src/main/java/com/oreocube/booksearch/feature/book/SearchBookViewModel.kt
@@ -13,8 +13,6 @@ import com.oreocube.booksearch.feature.book.model.BookUiState
 import com.oreocube.booksearch.feature.book.model.RecentHistoryUiState
 import com.oreocube.booksearch.feature.book.model.toUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -148,23 +146,4 @@ class SearchBookViewModel @Inject constructor(
             is SearchBookUiAction.ClearHistoryClick -> onClearHistoryClick()
         }
     }
-}
-
-data class SearchBookUiState(
-    val query: String = "",
-    val result: ImmutableList<BookUiState> = persistentListOf(),
-    val recentHistory: ImmutableList<RecentHistoryUiState> = persistentListOf(),
-)
-
-sealed class SearchBookUiEvent {
-    data class Error(val message: String) : SearchBookUiEvent()
-    data class NavigateToBookDetail(val isbn: String) : SearchBookUiEvent()
-}
-
-sealed class SearchBookUiAction {
-    data class InputChanged(val input: String = "") : SearchBookUiAction()
-    data class BookClicked(val book: BookUiState) : SearchBookUiAction()
-    data class HistoryItemClick(val history: RecentHistoryUiState) : SearchBookUiAction()
-    data class DeleteHistoryClick(val history: RecentHistoryUiState) : SearchBookUiAction()
-    data object ClearHistoryClick : SearchBookUiAction()
 }
